+++ src/main/java/com/greedy/mokkoji/api/notification/service/NotificationChannel.java
@@ -1,4 +1,4 @@
-package com.greedy.mokkoji.api.test.notification.service;
+package com.greedy.mokkoji.api.notification.service;
 
 import java.time.LocalDateTime;
 import java.util.List;
